requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.4
fsspec==2025.3.2
grad-cam==1.5.5
huggingface-hub==0.31.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
segmentation_models_pytorch==0.5.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
timm==1.0.15
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
ttach==0.0.3
typing_extensions==4.13.2
urllib3==2.4.0
waitress==3.0.2
Werkzeug==3.1.3
gunicorn==20.1.0
opencv-python-headless==4.8.1.78

